requirements.txt
Faker
geopy.distance
keyboard
numpy
pandas